Decrease max width of sessions table
Before this change, if the session length was anything other than
1 hour, it would be displayed as e.g. "45 minutes" or "1 hour,
30 minutes". This caused the sessions table to overlap the vertical
line separator. Now, the session length formatter function has changed
so that it would display "30 mins" or "1 hour 30 mins". In addition to
this, the width of the left column has been increased to 65% of the page
width and the width of the right column has been decreased to 35%.