Fix `print_inactive_students` function
After a major refactoring, I changed the `lessons` container from a
dictionary to a Pandas DataFrame but did not update this function. When
I generated invoices without the `--only` option (as I had been using
during my testing), it printed all the students in `data/students.json`,
even though I had seen most of them. This is now fixed.

I also fixed the `get_invoice_period` function. When I wrote this, I
intended for it to title invoices e.g. "Alice invoice for September"
if `start_date` and `end_date` were the start and end of the month, and
if not, to title them e.g. "Alice invoice for 05/04/2024 to 06/04/2025".
This was not working due (now fixed) faulty logic.